reorganize library code
Move library code to its own package, to allow for
easier importing from main module.

Add __init__ files to allow resolution, also from
outside (by tests).
