+++ src/__init__.py
@@ -1,0 +1,5 @@
+from . import nbp
+
+__all__ = [
+	"nbp"
+]
